+++ extendedSchemas.py
@@ -125,6 +125,7 @@ class SolutionDeploymentSchema(Schema):
     statusId = fields.Int()
     statusCode = fields.Str()
     statusMessage = fields.Str()
+    taskId = fields.Int()
 
 
 class ExtendedGoogleSessionSchema(Schema):
