--- mocks/mock_gcp_dac/extendedSchemas.py
@@ -1,13 +1,0 @@
-from marshmallow import Schema, fields
-
-class ResponseSchema(Schema):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-    id = fields.Int()
-    deployed = fields.Bool()
-    deploymentState = fields.Str()
-    statusId = fields.Int()
-    statusCode = fields.Str()
-    statusMessage = fields.Str()
-
