+++ mocks/mock_gcp_dac/gcpdac/extendedSchemas.py
@@ -6,3 +6,14 @@ class HealthSchema(Schema):
 
     status = fields.Str()
 
+class ResponseSchema(Schema):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    id = fields.Int()
+    deployed = fields.Bool()
+    deploymentState = fields.Str()
+    statusId = fields.Int()
+    statusCode = fields.Str()
+    statusMessage = fields.Str()
+    taskId = fields.Int()
